streaming_scraper/requirements.txt
-i https://pypi.org/simple
kafka-python==2.0.2
pyyaml==5.3.1
pafy==0.5.5
pymongo==3.11.0
youtube-dl==2020.12.2